m4 Common and others
Pulled what was variable strings into m4 files similar to what was done
in cppunit for the common framework stuff. Added a license directory in
m4 (doesn't really need to be m4 but whatever). Fixed some copyright
stuff (not done) and realized that the default configure.ac file needs
to be framework dependant (not sure why I forgot but here we are). Added
new testing functions to test out the newer features.