The too many edits comit
I've been working on this commit for so long at this point I don't even
remember what I was doing in the first place.  There was a lot of
consistance edits in terms of parameter order and parameters of
functions in general. Finished writing tests for ut-framework, made more
modifications to the m4 files. Created the ut.org to eventually be the
documentation file on github. Made changes to ut-cppunit-framework and
ut-common-framework in terms of the m4 changes.  More testing.