Framework plist reversion and framework testing
Move back from the plist macro in the ut-define-framework
definition. Even if it was a bit more concise in terms of code it was
more opaque in terms of what was going on there. Added documentation to
test-ut-framework because I kept falling down a rabbit hole in terms of
what to test. Now what will be test is layed out in the documentation
which should probably be the norm going forward (ie write *what* will be
tested before writing the tests)