Framework Testing Complete
Finished writing unit tests for framework and related functions. Pretty
sure i'm done twiddling with their definitions and what not for a
while now. Testing now relies on the ut-mock-framework functions instead
of defining the echo functions at the beginning. ut-mock-framework.el
will be where new mock functions are placed for further testing of UT.

ut-framework-*-hook functions now return the raw hook instead of the
value of the hook. There was some fall out and will be more from that
decision but I'm pretty sure that it is the correct one and is the more
expected behavior. Made changes in ut.el in terms of how the hooks are
being called as a result of the change.

ut-define-framework was pretty macro unfriendly with the
previous (unless cond (error )) in non-backquoted section. That has been
changed and can now happly live inside of other macros and not blow up
on expansion due to errors.