Renaming
Renamed a bunch of cppunit's m4 files to be a bit more descriptive. Even
I was getting confused which file was being use for what I'm the one
writing the fucking thing. Edits to the Makefile.am file in the src dir
of a test suite. Old one was pretty wrong and was missing a lot of
things that would be needed to run properly out of the box.