Parameter order cleanup
Cleaned up parameter ordering on every function I could identify in
ut.el and ut-cppunit-framework.el that had parameters out of
order. Ordering guidelines (because apparently I need to do this so that
I can write consistant code is `conf' is weighted the highest in terms
of appearing first, then `test-suite' then `test'.

A lot of function name clean up as well. Function names are now (mostly)
for the form: `ut-{structure}-{action}', so `ut-conf-project-name'
returns the project name from the conf structure and
`ut-test-suite-run' would run a test-suite.  Prior to this the naming
was all over the place.

Also in terms of consistency, prefer `new' to 'add', `del' over
`remove'. Need to still tackle predicate naming consistency, but not in
this patch.