Kill the buffer softly
Testing got a bit messy in parts because buffers created with
`with-current-buffer' were never being killed. This made some tests run
inconsistantly and succeed only after other tests created buffers with
the same name that it was planning on using.  Now all buffers are killed
at the end of each test.  Might toss that into a macro of some sort
later to make it a bit more tidy.

Also added new test to make sure that a blocking process in one project
space will not block a process in a different project space.