Killing the buffer even more slowly
Created a new macro to ensure that the ut sandbox is properly destroyed
after a test finishes (regardless of whether or not it succeeded.) This
was causing problems for tests that were being run in series (Run All)
where remnants of the previous test were either still alive or in the
case of some post functions were still running. Between the new
`with-ut-sandbox' macro and more liberal use of the
test-ut--sit-and-spin function's wait parameter all ut-process tests
succeed regeardless of what order they are run in.