A matter of time
Had to change up how some of the tests wait for certain autotools
processes to finish. Created a spin lock of sorts that checks to see if
the blocking process is done otherwise it waits for a second.  Needed to
make this change because the wait time will vary from computer to
computer, this is a bit safe.